Spin buttons deletion
spin buttons were not needed, are a distraction, and so a plain unadorned text
entry can be used instead.

Adding a Gtk.Entry widget to take input and performing a regular expression
match through a call back function "_value_insert_text" allowing only negative
or positive decimal valus and "." and ",".