Removing convert button and Updating POT file
Previously the convert button has to be clicked to calculate results.

Now on every change of value  the "_call" function will be called
and results will be calculated.
The Convert.pot file is also updated.
A placeholder is also placed instead of value "1" and check is introduced
for call of the _update_label() function to proceed only for integer values
to avoid float convert errors.

Fixes https://github.com/sugarlabs/convert/issues/16